Move ReservationSystem model to common app
